Fixed bug with GL gamma in GZDoom mode for some buggy drivers:
Fix for Radeon 9000, possibly other R200s: When the device is reset, it resets the gamma ramp, but the driver apparently keeps a cached copy of the ramp that it doesn't update, so when SetGammaRamp is called later to handle the NeedGammaUpdate flag, it doesn't do anything, because the gamma ramp is the same as the one passed in the last call, even though the visible gamma ramp actually has changed.

So here we force the gamma ramp to something absolutely horrible and trust that we will be able to properly set the gamma later.

git-svn-id: https://www.crowproductions.de/repos/prboom/branches/prboom-plus-24/prboom2@2743 df9a1a3e-e9e3-0310-a482-b5a01ba5a74e