Remove 16 extraneous bytes from CRBRICK
The CR* lumps are maps of Doom palette indices. CRRED is the identity
map. The others replace the red range 0xb0-0xbf with something else.

However in CRBRICK's case the 0xb0-0xbf mapping had been _inserted_ into
the lump, making all values above 0xbf map to the wrong colour, and
making the lump 16 bytes bigger than it should be.

The fix is to remove the 16 bytes 0xc0-0xcf (having values 0xb0-0xbf) so
that 0xc0, and everything above, map to themselves, as they should.