Revision: 3631 Author: rjy Date: 9:46:17, 3 листопада 2010 р. Message: configure.ac: add AC_GNU_SOURCE (Graham Gower)
Graham Gower reported on prboom-devel that I_SetAffinityMask was failing because sched.h does not export needed symbols without _GNU_SOURCE defined. I had not observed this bug because SDL puts -D_GNU_SOURCE=1 into CFLAGS on my system, but it is clearly a mistake to rely on this.

Graham's patch suggested to #define _GNU_SOURCE in i_system.c but I feel it's better for consistency and to pre-empt similar future problems by just setting it in config.h for the whole source tree.

git-svn-id: https://www.crowproductions.de/repos/prboom/branches/prboom-plus-24/prboom2@3645 df9a1a3e-e9e3-0310-a482-b5a01ba5a74e