P_LoadLineDefs: fix linedef numbers in error reports
Apparently in r2233 the error reporting moved from P_LoadLineDefs2 to
P_LoadLineDefs, but the number of the current linedef wasn't updated.

P_LoadLineDefs2 loops backwards and tests for zero:
  for (i = numlines, ld = lines; i--; ld++)
and the current linedef inside the loop is numlines-i-1.

But P_LoadLineDefs loops over the lines array forwards i.e.
  for (i = 0; i < numlines; i++)
so the current linedef inside the loop is just i.

Thanks to Never_Again for his bug report on Speed of Doom map23,
in which the linedef numbers appeared on examination to be bogus.
http://www.doomworld.com/vb/post/932261